created all model

+++ src/main/java/com/albumbazaar/albumbazar/dao/LocationRepository.java
@@ -4,6 +4,6 @@ import com.albumbazaar.albumbazar.model.Location;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocationRepository extends JpaRepository<Location, Integer> {
+public interface LocationRepository extends JpaRepository<Location, Long> {
     
 }
